+++ R/RcppExports.R
@@ -697,6 +697,29 @@ rxMvrandn_ <- function(A_, mu, sigma, lower, upper, ncores = 1L, a = 0.4, tol = 
     .Call(`_RxODE_rxMvrandn_`, A_, mu, sigma, lower, upper, ncores, a, tol, nlTol, nlMaxiter)
 }
 
+rxSeedEng <- function(ncores = 1L) {
+    .Call(`_RxODE_rxSeedEng`, ncores)
+}
+
+rxnorm_ <- function(mean, sd, n, ncores) {
+    .Call(`_RxODE_rxnorm_`, mean, sd, n, ncores)
+}
+
+#' Simulate normal variable using the sitmo vandercorput engine
+#' 
+#' @inheritParams rxnorm
+#'
+#' Note the vandercorput is not good random number generator, rather
+#'   it is a low discrepancy random number generator which covers the
+#'   random surface better than a good random number generator
+#'   should.  This can actually help estimation methods relying on
+#'   random number generation, but is not good for simulation.
+#' 
+#' @export
+rxnormV_ <- function(n, mean, sd, ncores) {
+    .Call(`_RxODE_rxnormV_`, n, mean, sd, ncores)
+}
+
 # Register entry points for exported C++ functions
 methods::setLoadAction(function(ns) {
     .Call('_RxODE_RcppExport_registerCCallable', PACKAGE = 'RxODE')